+++ list-tx.go
@@ -50,12 +50,12 @@ func (tx *ListTx) Count() (int64, error) {
 	return res.(int64), nil
 }
 
-func (tx *ListTx) Size(index int64) (int, error) {
+func (tx *ListTx) Size(index int64) (int64, error) {
 	o, err := tx.Client().SQLQueryFirst("select coalesce(length(value), 0) from $table$ where ind=?;", IntScanner, index)
 	if err != nil {
 		return 0, err
 	}
-	return o.(int), nil
+	return o.(int64), nil
 }
 
 func (tx *ListTx) TotalSize() (int64, error) {
